Add persistent SNP arc highlighting to LD track
Introduces a 'show' prop and config option to allow users to persistently display LD arcs for selected SNPs in the LD track. Clicking a SNP toggles its presence in the 'show' list, and arcs/labels are rendered for all SNPs in this list as well as the currently hovered SNP. Updates types to support the new 'show' property.